# Deduplicate project remote dependencies