--- a/src/Juvix/Compiler/Pipeline/Loader/PathResolver/Data.hs
+++ b/src/Juvix/Compiler/Pipeline/Loader/PathResolver/Data.hs
@@ -7,6 +7,7 @@ import Juvix.Prelude
 
 data ResolverEnv = ResolverEnv
   { _envRoot :: Path Abs Dir,
+    _envProjectRoot :: Path Abs Dir,
     -- | The path of the input file *if* we are using the global project
     _envSingleFile :: Maybe (Path Abs File),
     _envLockfileInfo :: Maybe LockfileInfo
@@ -60,6 +61,9 @@ checkRemoteDependency d = case d ^. resolvedDependencyDependency of
 
 withEnvRoot :: (Members '[Reader ResolverEnv] r) => Path Abs Dir -> Sem r a -> Sem r a
 withEnvRoot root' = local (set envRoot root')
+
+withEnvProjectRoot :: (Members '[Reader ResolverEnv] r) => Path Abs Dir -> Sem r a -> Sem r a
+withEnvProjectRoot projectRoot = local (set envProjectRoot projectRoot) . local (set envRoot projectRoot)
 
 withLockfile :: (Members '[Reader ResolverEnv] r) => LockfileInfo -> Sem r a -> Sem r a
 withLockfile f = local (set envLockfileInfo (Just f))