--- a/src/Juvix/Compiler/Pipeline/Loader/PathResolver/Data.hs
+++ b/src/Juvix/Compiler/Pipeline/Loader/PathResolver/Data.hs
@@ -6,8 +6,10 @@ import Juvix.Compiler.Pipeline.Package.Base
 import Juvix.Prelude
 
 data ResolverEnv = ResolverEnv
-  { _envRoot :: Path Abs Dir,
-    _envProjectRoot :: Path Abs Dir,
+  { -- | The root path of the current project being resolved
+    _envRoot :: Path Abs Dir,
+    -- | The root path of the initial project (i.e the first project considered in the resolution)
+    _envInitialRoot :: Path Abs Dir,
     -- | The path of the input file *if* we are using the global project
     _envSingleFile :: Maybe (Path Abs File),
     _envLockfileInfo :: Maybe LockfileInfo
@@ -62,8 +64,8 @@ checkRemoteDependency d = case d ^. resolvedDependencyDependency of
 withEnvRoot :: (Members '[Reader ResolverEnv] r) => Path Abs Dir -> Sem r a -> Sem r a
 withEnvRoot root' = local (set envRoot root')
 
-withEnvProjectRoot :: (Members '[Reader ResolverEnv] r) => Path Abs Dir -> Sem r a -> Sem r a
-withEnvProjectRoot projectRoot = local (set envProjectRoot projectRoot) . local (set envRoot projectRoot)
+withEnvInitialRoot :: (Members '[Reader ResolverEnv] r) => Path Abs Dir -> Sem r a -> Sem r a
+withEnvInitialRoot projectRoot = local (set envInitialRoot projectRoot) . local (set envRoot projectRoot)
 
 withLockfile :: (Members '[Reader ResolverEnv] r) => LockfileInfo -> Sem r a -> Sem r a
 withLockfile f = local (set envLockfileInfo (Just f))