--- a/src/Juvix/Compiler/Pipeline/Loader/PathResolver/DependencyResolver.hs
+++ b/src/Juvix/Compiler/Pipeline/Loader/PathResolver/DependencyResolver.hs
@@ -36,7 +36,7 @@ runDependencyResolver = runProvider_ helper
                   _resolvedDependencyDependency = i ^. packageDepdendencyInfoDependency
                 }
           DependencyGit g -> do
-            let r = rootBuildDir (env ^. envProjectRoot)
+            let r = rootBuildDir (env ^. envInitialRoot)
             gitCacheDir <- globalGitCache
             let cloneRelDir :: Path Rel Dir
                 cloneRelDir = mkSafeDir (g ^. gitDependencyUrl)