… tests/smoke/Commands/compile-dependencies-package-juvix.smoke.yaml
… tests/smoke/Commands/compile-dependencies.smoke.yaml